Webelement Location is now eager
Prior to this, finding WebElements was done lazily, when the element was actualy used.
While this had some elegance it proved to add problems:
The elements function, which returns a list of elements
returned a list, which might change by the point the elements where actually used.
In particular the visibility check within WebElement might test a different selenium element
than the one being wrapped by the WebElement.