Add finalizer which closes window to scope functions.
Fixes a bug where an exception during a browse() call
would cause Browser to stop without calling close() on the current window.
Cadmium would thus leak WebDriver instances.
This would be caused by failing assertions in unit tests for example.

Add a hook which is called when an exception is thrown during a browse() session.